+++ js/data.js
@@ -1,0 +1,13 @@
+const SIMILAR_DESCRIPTIONS_COUNT = 25;
+const PHOTO_DESCRIPTIONS = ['Пляж отеля', 'Go to the beach', 'Цвет моря', 'Подруга', 'Супчик', 'Макларен', 'Десерт', 'Вкусный напиток', 'Летают самолеты', 'Храним обувь', 'Путь к морю', 'Ауди', 'Ужин', 'Пункт меню', 'Отдыхаю', 'Лечу в отпуск', 'Концерт', 'Ретро-машина', 'Тапочки с подсветкой', 'Пальмы', 'Вилка', 'Закат', 'Краб', 'Руки на концерте', 'Бегемот',];
+const MIN_LIKES_COUNT = 15;
+const MAX_LIKES_COUNT = 200;
+const MIN_ID_DESCRIPTIONS_COUNT = 1;
+const MAX_ID_DESCRIPTIONS_COUNT = 25;
+
+const MIN_COMMENT_COUNT = 1;
+const MAX_COMMENT_COUNT = 4;
+const USERS_NAMES = ['Дарья', 'Алексей', 'Иван', 'Наталья', 'Юлия', 'Андрей',];
+const COMMENTS_VALUE = ['Всё отлично!', 'В целом всё неплохо. Но не всё.', 'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.', 'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',];
+
+export {SIMILAR_DESCRIPTIONS_COUNT, PHOTO_DESCRIPTIONS, MIN_LIKES_COUNT, MAX_LIKES_COUNT, MIN_ID_DESCRIPTIONS_COUNT, MAX_ID_DESCRIPTIONS_COUNT, MIN_COMMENT_COUNT, MAX_COMMENT_COUNT, USERS_NAMES, COMMENTS_VALUE};
